CR-998 move Firestore classes and tests from RHSvc to framework
