+++ pom.xml
@@ -96,7 +96,7 @@
     <dependency>
       <groupId>uk.gov.ons.ctp.integration.common</groupId>
       <artifactId>framework</artifactId>
-      <version>0.0.59</version>
+      <version>ROBC-CR-998-SNAPSHOT</version>
     </dependency>
 
     <dependency>
